requirements.txt
minio
roboflow
python-dotenv
kedro
confluent-kafka
